Complete SoC of view and functionality + refactor of Image component
