+++ packages/plugins-image/src/cadmus-components/imageRendererHelper.js
