+++ packages/plugins-image/src/schemas/index.js
@@ -7,7 +7,6 @@ const schema = {
     normalize: (change, reason, { node, child }) => {
       switch (reason) {
         case LAST_CHILD_TYPE_INVALID: {
-          console.log(LAST_CHILD_TYPE_INVALID);
           const paragraph = Block.create("paragraph");
           return change.insertNodeByKey(node.key, node.nodes.size, paragraph);
         }
